requirements.txt
google-genai>=0.2.0
google-adk>=0.1.0